Bug #19651687 : THE SHA256 CLIENT PLUGIN CAN\'T BE USED AS A DEFAULT PLUGIN
The sha256 client plugin was resetting the password into the connection handle.
Thus, if used first in a 2 client-plugin authentication exchange, it was 
depriving the second client plugin from access to the password supplied.
Fixed by re-moving the reset of the password copy (to support MYSQL_OPT_RECONNECT)
No need for a test, since this problem was discovered by an existing test.

Also removed an unused paramer compilation warning when compiling non-debug.