Fix Bug#19472028 BUG#18663997 WAS NOT (FULLY) FIXED, STILL: CANNOT CLOSE FILE: MOD CNT!=FLUSH CNT
Remove the mtr testcase import_flush_temp.test and the changes in the code
that were made solely for this test. The test is too aggressive in trying
to reproduce Bug#18663997 that even the error printout is emitted for tables
like mysql.innodb_table_stats that are not temporary and have never been
imported.

Approved by:	Marko (via IM)