Patch 2 for WL#7953 Transporter handshake retry
Change the server side of Transporter/TransporterRegistry to request the client to close() the
socket first when server is not yet ready to accept the connection. This moves the TIME_WAIT
state for the socket to the client and thus frees resources  from the network stack on the server.

Requesting client to close is done by sending garbage to the client. Although we send the nice text
string BYE to the client it will be interpreted as garbage and it will close the connection. After
sending the BYE string wait for the EOF on the socket by reading 0 bytes from socket. ie. select
says there is something to read but you only read 0 -> EOF.  Since connections are handskaed serially
wait for EOF with fairly short timeout in case there should be a connected client which does not
behave as expected. After the new BYE and wait for EOF code the server side socket is closed as
usual and the TIME_WAIT TCB shold now end up on the client where the previous "use last port" patch
will eliminate it.

Also change so that socket is closed with a reset(RST) when the handshake is rejected
due to invalid parameters or other error. This effectively eliminates TIME_WAIT TCB from both
server and client side for the non graceful case(s).