Bug#19685095 DO NOT CARE ABOUT UNRESOLVED MLOG_FILE_NAME IF THERE ARE NO OPERATIONS TO APPLY
The MLOG_FILE_NAME record that was introduced in WL#7142 can be
emitted on a log checkpoint even if the file was not changed.

If the file is missing or unreadable on crash recovery (redo log
apply), we should ignore this inconsistency unless there actually is
some redo log to apply on the tablespace.

rb#6280 approved by Kevin Lewis and Mattias Jonsson