Bug#18487960: SIG 11 on delete with index merge
The mysqld cores while attempting read removal before a delete with 
an index merge. The core occurs only if a quick select is generated 
for the delete operation.

During read removal, the index from the quick select is used to index 
into the table structure. But since the delete is an index merge, the 
quick select index is set to MAX_KEY instead of a valid index value. So 
the indexing operation generates a garbage pointer, which is then 
dereferenced, resulting in a core.

Added a check so that read removal before delete is not attempted if 
the quick select index has a value of MAX_KEY.