Bug #17257842 EXCESSIVE RETRY FREQUENCY FOR API NODES RECONNECT CAUSES HIGH CPU ON MGM NODES.
For system that have lot of API nodes and also have added data node
entries in configuration without starting any process for them the
API nodes will continously try to connect to the missing data nodes
several times per second.  This put extra load on MGM nodes and
network in general.



This patch introduces logic to make an API node wait longer and
longer between attempts to connect to data nodes that fail to
respond.

The logic is controlled by two new parameters for api nodes:
StartingConnectingBackoffMaxTime and ConnectingBackoffMaxTime.

Both are given in milliseconds with approximately 100ms resolution.
Note that the time controls time *between* attempts, during the
time a connect attempt occur no time float with respect to the
backoff logic.

Between first and second attempt one waits 100ms (as before), but
for the following attempts the wait is doubled until the configures
maximal time are reached or ~100s whichever is smallest.

As long as the API node is not connected to any data node the
StartingConnectingBackoffMaxTime parameter is used, otherwise
ConnectingBackoffMaxTime.

Note, as soon as the API node is connected to a data node and that
node reports, in heartbeat message, that it has connected to some
other data nodes, connection attempts to those data nodes will not
be affected by the backoff logic but will be attempted every 100ms
until connected.

In worst case it can take HeartbeatIntervalDbApi (default 1,5s)
until the API node get notified that a restarting data node is up.

But since data nodes connects to each other during start phase 2,
and API node connections are allowed from start phase 8, that
should normally not be an actual delay since the time between start
phase 2 and 8 probably are at least several seconds.

For systems with many unstarted data nodes the new parameters
should be raised to avoid extreme traffic either to MGM nodes or
connection attempts to non existing datanodes.