Bug#19050141 SHOW EVENTS IN PERFORMANCE_SCHEMA GIVES ACCESS DENIED
Before this fix, SHOW EVENTS in the performance_schema database
returned an access denied error.

Technically, the error is actually expected, because no user
can be granted the EVENT_ACL privilege on the performance schema.

From an ease of use point of view, 
and considering that an exception for the INFORMATION_SCHEMA
is already implemented, it is desirable to relax the check,
and return an empty result set instead of an access denied error
in this case.

This fix implements an exception for the PERFORMANCE_SCHEMA
in SHOW EVENTS, for consistency with existing behavior.