WL#8109-InnoDB: Refactor code for general tablespaces
This is a group of changes related to how tablespaces are used.
They do not affect any output or test cases. They are completely
non-functional.

1) Refactor code that initializes, validates and handles tablespace
flags and table flags. This is done so that a SHARED and TEMPORARY
flag can be added to FSP_FLAGS by wl6205.
2) Rename functions and variables to refer less to single-table
tablespaces when they will be used in wl6205 for general tablespaces
as well. Add "_for_table" to 'dict_build_tablespace' since it will be
specifically for file-per-table.
3) Refactor fil_create_ibd_tablespace to allocate and free the filepath
in the caller, and other cleanup.
4) Add a few asserts that are always true.
5) Other whitespace and comment cleanups.

Approved by Vasil in RB#6669