Make a fault injection test more robust.
ULINT32_UNDEFINED will match ULINT_UNDEFINED on 32-bit systems.
That is the initial space id of srv_tmp_space, before it is assigned.

Also, skip the test if a redo log checkpoint occurred,
because it could fail in that case. We are testing the
redo log scan, and we do want to see our extra record
after the latest checkpoint.