WL#8109 Follow-up
The previous patch was a non-functional change extracted from wl#6205.
But it did one thing which exposed problems in page_size=4k & 8k runs.
That is, it made a requirement that page_size_t objects always use
non-zero page sizes. This is a much better approach becasue zero was
getting interpreted by the code that uses page_size_t, sometimes to
UNIV_PAGE_SIZE (which can be 4k,8k,16k,etc) and sometimes to
UNIV_PAGE_SIZE_ORIG (16k). The original patch had a bug which caused
all 4k runs to fail which was fixed quickly by Marko. That left only
innodb_double_write.test failing which this patch fixes.

The innodb_double_write test writes all zeros to the header page of the
system tablespace, hoping that the double-write recovery can fix it.

This patch (a) changes 2 places where a page_size_t is initialized with
zero to use non-zero numbers. (b) It stops initializing system tablespaces
from the contents of the first page, which may be corrupted, like it is in
innodb-double-write.test. It uses the space_id and flags that system
tablesapces are known to use. (c) It moves the assert that physical() is
not zero when UNIV_PAGE_SIZE is not 16k to the function that puts the page
size into an FSP flags.  This means we can have an invalid page_size_t and
an FSP flags read from a corrupted file, but we are not going to build an
invalid FSP flags.  

approved by Marko in rb#6897