WL#8109 changed in page_size_t to enforce the use of non-zero page sizes. But some lingering affects have been noticed. It was found that the test innodb_page_size_func.test fails when using innodb-page-size=8k, because the expected error message was not seen.
This patch switches the order of error checking in
Datafile::validate_first_page() so that the most appropriate and
informative log message is given, which is the message that the test
expects.

There are a couple other non-functional cleanup changes in this patch
as well.

Approved in rb#6914 by Marko