Bug#19730970: COMPILE ERROR WITH ENABLED_PROFILING=0
This patch fixes two compilation errors when building with 
-DENABLED_PROFILING=0.

The first is the missing inclusion of sys/resource.h in mysqld.cc which otherwise
gets included via sql/sql_profile.h. This was a regression introduced in 5.7.5 by
the patch for Bug#18404381.

The second is a missing field.h inclusion in the thread pool.