Bug #19711674 SETTING REPLICATE_WILD_IGNORE_TABLE FILTER CAUSES CRASH
Problem: Setting replication_[wild]_[do|ignore]_table to an empty
value is causing server to go down.

Analysis:
In the server, a check is there to cleanup [do|ignore]table_hash
if number of records in that hash are empty. But this check
is not protected with whether the hash initalized or not.
Hence in this situation, it tries to delete some garbage
hash buffer which is resulting in above said problem..

Fix:
Now the check ([do|ignore]_table_hash_inited == true)
is added before the server tries to free the hash.