BUG#14798022 : "CHAR (0)" DATATYPE CAUSING ERROR 1296 (HY000) : GOT ERROR 4547 : 'RecordSpecification has overlapping offsets'
CHAR(0) is internally represented as BIT(1) by the NDB. MySQL assigns 1
null bit for the CHAR(0) column and assigns the next bit to the next column's
null bit. But as the NDB treats CHAR(0) as BIT(1), it uses 2 bits in the null
bitmap - one for data and one for the actual null bit. Hence the overlap.
This patch fixes that by making the NDB aware of the CHAR(0) column.
changes:
 -added a column_flag to the RecordSpecification struct
  to pass per column flags to NDB
 -modified the validateRecordSpec and bitfield handlers to skip reading 
  the overflow bits when there is a CHAR(0) 
 -older version of RecordSpecification has been maintained for backward
  compatabililty reasons
 -moved the logic from createRecord to createRecordInternal
 -createRecord, if gets an older version of RecordSpec, maps it to a newer 
  version and calls createRecordInternal internally.
 -added various test cases for char(0) and RecordSpecification