Bug #19582925 ENHANCED DETECTION AND DUMP OF CORRUPT OR UNSUPPORTED MESSAGES
Add more checks while unpacking received signals over tcp.

Byte order must be the native byte order.
Compressed flag may not be used.
Check also byte order, compressed flags, and message length for
next message if in receive buffer.

If either current unpacked message or next fails in check, current
message are assumed bad.

If message is bad, the transporter are marked having bad data and
no more unpacking of messages will occur until reconnected.

When bad message is detected a timestamped log with error and
hexdump of message and if next message was bad also a hexdump of
first part of that and the six preceding words.