Bug#19471564 WL#6737:IN-CONSISTENT DATA-TYPE FOR GEOMETRY TYPES WHILE PROJECTING DERIVED TBL.
Problem: A union query where the first query block returns a POINT
column and the second returns a geometric column with a non-point
value will fail with an error message if the query is using InnoDB
temporary tables or storing the result in an InnoDB table.

The column types of a union query are the combinations of the column
types of the query blocks of the query. In
st_select_lex_unit::prepare(), the column types of each query block
are inspected in the order the query blocks appear in the query.

The combined column types start out as the types of the first query
block. Each type is stored in a Item_type_holder object, which not
only stores the field type (which is common to all geometric column
types), but also the geometry type.

For each column of each subsequent query block,
st_select_lex_unit::prepare() calls Item_type_holder::join_types() to
merge that type into the combined type of this column in previous
query blocks. This function correctly infers the combined field type,
but it neglects to combine the geometry types, leading to the column
being of the geometry type of the column in the first query block.

If the first query block has a POINT column, the combined type will be
a POINT column, which in InnoDB can't store other geometric
types. Therefore, the insertion of such objects from a subsequent
query block will fail.

Fix: In Item_type_holder::join_types(), if the current combined
geometry type and the geometry type of the column in the new query
block is different, use a GEOMETRY column.