BUG#19667258: WL#6972: SERVER CRASH ON SET SESSION SESSION_TRACK_GTIDS='OFF'
The server crashes when changing the value of session_track_gtids from
'ALL_GTIDS' or 'OWN_GTID' to 'OFF'. This was due to the fact that the
memory structures were being released to early in the process.

We fix this by deferring the internal memory structures cleanup until
after the OK packet has been filled with the required information.

Furthermore, this patch fixes the behavior regarding the GTID
information that is put in the OK packet during transitions between
states of session_track_gtids, i.e., when the user issues 
 
  SET session_track_gtids='...'.

The behavior is the following during transitions (from -> to):

  |-----------+-----------------------------------|
  |    FROM   |                TO                 |
  |-----------+-----------+-----------+-----------|
  |           | OFF       | OWN_GTID  | ALL_GTIDS |
  |-----------+-----------+-----------+-----------|
  | OFF       | No GTID   | No GTID   | No GTID   |
  | OWN_GTID  | No GTID   | No GTID   | No GTID   |
  | ALL_GTIDS | ALL_GTIDS | ALL_GTIDS | ALL_GTIDS |
  |-----------+-----------+-----------+-----------|

This patch also includes tests for the various transitions depicted in
the above table.