BUG#19815883: WL 6128 - ASSERTION ON REGISTER_CTX_CHANGE_LISTENER
Fixed the case that the connection is cleaned up and the tracker
destroyed, but the listener would not be unregistered from the
GTIDs collector. This could happen when calling change user and
might later raise an assertion. The assertion was due to the fact
that a new listener would be registered in the GTIDs collector while
the reference to the old listener was still there. This triggers the
assertion.

The fix is to call the GTIDs collector unregister function when the
tracker is destroyed.

In addition, restricted GTID tracker activities to NON_SYSTEM_THREADs.