Non-functional changes extracted from wl6205 - Part 11
Rename the functions that update the data dictionary from 'insert' to
'replace' and add SQL to delete the previous SYS_TABLESPACES and
SYS_DATAFILES records before inserting new records. This is done because
this action is taken when the dictionary needs to store a tablespace
record plus all its known tablespace files. It is really an UPDATE in
the sense that it should replace what is there, if anything, with the
correct metadata. If there is a stray SYS_DATAFILES record with this
space_id, then it will no longer remain. I don't think this affects any
current conditions, but it makes a cleaner interface with the data
dictionary.

Approved by Vasil in rb#7003