Bug #18817867 MYSQLDUMP DOES NOT REPORT ERROR 28 WHEN WRITING TO FULL NFS MOUNT
Description:
While taking mysqldump on full NFS mount location it does not report an error(errno 28) like it should report on a full local disk.

Analysis:
At the end of mysqldump we are calling fflush() to flush the stdout data to disk file. If the stdout file is associated to  local disk then it was throwing error 28(disk full error). If the stdout associated to nfs file location then fflush() is putting the date in OS I/O buffers and not reporting the error. Hence we are not getting error.

Fix: As fix we are making sure to flush the data to the disk file by calling fsync(). Since the disk is full at nfs mounted path fsync() fails and we get error(errno 5; means write I/O error).

Note: This cause mysqldump performance impact. But we can guarantee that the data is flushed successfully to the disk file.