BUG#19368904 - INNODB GIS: ALTER TABLE AFTER UPGRADE FROM 5.6 CRASHES THE SERVER
In mysql 5.6, we store all GIS types as DATA_BLOB in InnoDB layer, also
in mysql 5.6, we don't allow users to create SPATIAL indexes on
GIS columns. But in 5.7, we change most the types to DATA_GEOMETRY and
introduce the spatial indexes. So there could be some issues when we
try to create spatial indexes on old 5.6 tables.

We will update the mtypes of all legacy GIS columns from DATA_BLOB to
DATA_GEOMETRY when users want to create spatial indexes on them.
There are 2 cases in this solution: 1) If we have to rebuild the whole
table, all old GIS columns would be updated to DATA_GEOMETRY;
2) If we can create spatial indexes in-place, then only the columns
on which we will create spatial indexes would be updated.

rb#6758, approved by Jimmy, Marko