Bug #19297190 	NOT IN DOESN'T RETURN EXPECTED RESULT 5.7-specific fix. A query uses subquery materialization. The inner expression is varchar(500), which is too long for indexing, so the temporary table is created with a "unique constraint" (hash_field), and ha_index_init() is done for the indexed hash_field. When the tmp table is filled, row after row, duplicates are eliminated by checking them with check_unique_constraint() which uses ha_index_read_map(). If there is a lot of data this MEMORY table becomes full and create_ondisk_from_heap() changes it to MyISAM or InnoDB; however this last function does ha_index_or_rnd_end(), then reads the MEMORY table with ha_rnd* (to fill the disk-based table); when this function ends, the index of the new on-disk table is not initialized for reading. So when we try to insert a next row, check_unique_constraint() runs and ha_index_read_map() raises an assertion because the index is not ready; this holds for trunk. In 5.6, a bad result is returned. The difference in behaviour is due to a lot of refactoring of internal tmp tables done by wl#6711.
This problem exists only if the index is for hash_field; if a regular
index is used for uniqueness checking it is checked inside the engine,
so does not require ha_index_init.
Those features use index-over-hash-field:
- subquery materialization: affected
- UNION DISTINCT: affected (uses same code as subquery
materialization: select_union::send_data()).
- semijoin materialization lookup: not affected, because if hash_field
is used this strategy censors itself and goes to materialization-scan
(well, this is going to be true when we push the last, patch-pending
fix for Bug 18227171)
- semijoin duplicates weedout: not affected, because if hash_field is
used then table is directly created as on-disk
- GROUP BY, DISTINCT, ROLLUP: not affected, because after every write
to the tmp table, QEP_tmp_table::prepare_tmp_table() is called, which
does ha_index_init.

Fix: select_union already initializes the hash_field index after
creating the table, let is initialize it again after the table has
changed engine.