Bug#19657747 STACK BUFFER OVERFLOW WITH GEOJSON FUNCTIONS
Description:
When ST_AsGeoJson() was given an illegal max_dec_digits or options argument that would evaluate to a number with length MAX_BIGINT_WIDTH (20 characters), a stack buffer overflow would occur. This is because a null terminator '\0' would be appended when converting this number to a string, which it wasn't allocated space for.

Fix:
When intializing the char array to hold the string representation of the illegal number, allocate one extra byte for the null terminator '\0'. Extra tests were also added to ensure that this won't happen for ST_GeomFromGeoJson() as well. This patch also fixes correct output of error messages when large unsigned integers are given as argument.