Bug #19438612: HANDLING OF DEPRECATED --SKIP-SECURE-AUTH OPTION PROBLEMATIC
Fix #1:
Made --secure-auth behave like this:
1. When not specified keep quiet about it.
2. When specified with the default value, issue a warning it's deprecated
and go on.
3. When specified with a non-default value issue an error.

Fix #2: Before printing the help text for the mysql options turn off
the default for secure-auth to prevent the my_getopt code from 
printing the "on by default, use --skip-secure-auth" text. 
This is safe to do as the help list of options is followed by a call to exit()

Distributed the fix to the rest of the client utils with the correct utility
name for each.