Fixes restart race causing test timeout in AutoTest 'testDict -n schemaTrans'
The test cases v1,v2, and x1,x2 restarted a node without the 
'NoStart' option, and then waited for the node to reach the
'NO START' state. However, this opens the possibility of
the node to restart, and reach the 'STARTING' state without the
test client ever seing the 'NO START' state, Thus the test
would finaly time out (Fail) after 120s.