Fix for bug#19712569
    NODE RESTART BLOCKED BY DICT LOCK HELD BY SCHEMA TRANSACTION
      
The dictionary API code may incorrectly interpret a TRANS_END_REPORT
belonging to a previous TRANS_BEGIN_REQ as a failure of the latest
schema transaction. This caused the API to assume that the transaction
failed and its locks was released, while the transaction was actually
started. This transaction is then effectively lost, and its locks
prevents other schema transactions to be started. As node restarts also
need the DICT lock, they are also prevented from getting past start phase 2.

This fix ignores TRANS_END_REP signals which belongs to unknown Txn,
or to transactions not yet started. To correctly track the Txn state,
we now also defer setting Txn-state to 'Started' from sending
the TRANS_BEGIN_REQ until TRANS_BEGIN_CONF is received.

Furthermore, this patch also changes the expected outcome of starting a 
schema transaction during a master node failure: If a master node now fails
before a TRANS_START_CONF has been sent, ::dictSignal() will resend
the TRANS_START_REQ signal to the new master, and it will
likely succeed the next time. (Which it actually also did before,
but the API incorrectly assumed that it failed)