WL#7440: Moving binlog event decoding into separate package
Fixing bug in scripts/CMakeLists.txt that causes linktest to fail. The reason
for the failure is that the code only take into account the target name, and
not the OUTPUT_NAME property of the target, which could be set to something
else.

The patch fixes the issue by checking if there is a value for the OUTPUT_NAME
target property and using that if there is, otherwise, it uses the name of the
target as the output name.