Bug#19670163 : ASSERT AFTER MISSING ERROR CHECK IN SUBQUERY
DESCRIPTION:

If a DML statement containing a subquery caused a deadlock inside
InnoDB, InnoDB would roll back the transaction. This would not be
noticed in the SQL layer where the execution would continue
eventually leading to an assert inside InnoDB.

FIX:

Added check for error. Abort statement on error condition.