Bug #19824442	ALTER TABLE FAILS, TABLESPACE ALREADY IN CACHE, GENERIC ERROR FROM ENGINE
The dict_table_t::flags2 is not read from disk for row format redundant because
we don't trust the MIX_LEN field in the system tables. Change the if condition
introduced by revid: kevin.lewis@oracle.com-20141010161318-tawwqdoo1cgwvind
so that we don't use the dict_table_t::flags2 but use the system ID to check
if it is table that lives in its own tablespace.

Approved by Jimmy Yang over IM.